\name{ofw-internal}
\alias{learnCART}
\alias{learnCART.default} 
\alias{learnSVM}
\alias{learnSVM.default} 
\alias{evaluateCART}
\alias{evaluateCART.learnCART}
\alias{evaluateSVM}
\alias{evaluateSVM.learnSVM}
\alias{ofwCART}
\alias{ofwCART.default} 
\alias{ofwSVM}
\alias{ofwSVM.default}
\alias{simplexe}
\alias{svmlearn}



\title{Internal ofw functions}
\description{Internal functions in ofw: \code{learnCART}, \code{learnSVM}, \code{evaluateCART.learnCART}, 
\code{evaluateSVM.learnSVM}, \code{ofwCART}, \code{ofwSVM}, \code{simplexe}, \code{svmlearn}.}


\details{
These functions are not to be called by the user.
}
	

\author{Kim-Anh L\^e Cao \email{Kim-Anh.Le-Cao@toulouse.inra.fr}

Patrick Chabrier \email{Patrick.Chabrier@toulouse.inra.fr}}


\keyword{classif}% at least one, from doc/KEYWORDS
%\keyword{tree}


