\name{srbct}
\docType{data}
\alias{srbct}
\alias{class}



\title{Small version of the small round blue cell tumors of childhood data for illustration purpose}
\description{
  This data set gives the expression measure of a small subset of genes that were randomly sampled from the the original data set from Khan et al. 2001. 
  }
}

\usage{
data(srbct)}

\format{
	A data frame \code{srbct} containing 200 genes and 63 microarrays (observations).
	A vector \code{class} containing the class tumour of each observation.
}
        

\source{http://research.nhgri.nih.gov/microarray/Supplement}
\references{Khan et al. (2001) Classification and diagnostic prediction of cancers using gene expression profiling and artificial neural networks. {\it Nature Medicine} {\bf 7} Number 6, June.}

\keyword{datasets} % at least one, from doc/KEYWORDS
%\keyword{microarray} a mettre mais j'ai pas la liste des kwords


