\name{ofw-internal}
\alias{learnCART}
\alias{learnCART.default} 
\alias{learnSVM}
\alias{learnSVM.default} 
\alias{evaluateCART}
\alias{evaluateCART.learnCART}
\alias{evaluateSVM}
\alias{evaluateSVM.learnSVM}
\alias{ofwCART}
\alias{ofwCART.default} 
\alias{ofwSVM}
\alias{ofwSVM.default}



\title{Internal ofw functions}
\description{Internal functions in ofw: \code{learnCART}, \code{learnSVM}, \code{evaluateCART.learnCART}, \code{evaluateSVM.learnSVM}, \code{ofwCART}, \code{ofwSVM}.}

%\usage{
%\method{learnCART}{default}(x, y, ntree=50, nforest=10, mtry=5, do.trace=FALSE, nstable=50, weight=FALSE, Bsample=5, ...)
%\method{learnSVM}{default}(x, y, niteration=1000, mtry=10, do.trace=FALSE, nstable=50, weight=FALSE, Bsample=5,...)
%\method{evaluateCART}{learnCART}(obj, maxvar=15, nvar=obj$nclass+1, ntreeTest=100, weight=FALSE, ...)
%\method{evaluateSVM}{learnSVM}(obj, maxvar=15, weight=F, ...)
%\method{ofwCART}{default}(x, y, ntree=50, nforest=100, mtry=5, do.trace=FALSE, nstable=25, keep.inbag=FALSE, keep.forest=TRUE, weight=FALSE, ...)
%\method{ofwSVM}{default}(x, y, niteration=100, mtry=5, do.trace=FALSE, nstable=25, weight=FALSE, ...) 
%}
	    
	   
%\arguments{
%}

\details{
These functions are not to be called by the user.
}
	
%\value{}

%\note{}

%\references{}

\author{Kim-Anh L\^e Cao \email{Kim-Anh.Le-Cao@toulouse.inra.fr} \newline
Patrick Chabrier \email{Patrick.Chabrier@toulouse.inra.fr}}

%\seealso{}

%\examples{}
\keyword{classif}% at least one, from doc/KEYWORDS
%\keyword{tree}


